requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alabaster==1.0.0
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anthropic==0.57.1
anyio==4.9.0
arabic-reshaper==3.0.0
arrow==1.3.0
asn1crypto==1.5.1
asttokens==3.0.0
asyncio-throttle==1.0.2
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
backports.tarfile==1.2.0
backrefs==5.9
bcrypt==4.3.0
beautifulsoup4==4.13.4
binaryornot==0.4.4
black==25.1.0
blinker==1.9.0
blis==1.3.0
build==1.2.2.post1
CacheControl==0.14.3
cachetools==5.5.2
cadquery==2.5.2
cadquery-ocp==7.7.2
casadi==3.7.0
catalogue==2.0.10
certifi==2025.7.14
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.2
chromadb==1.0.15
cleo==2.1.0
click==8.2.1
cloudpathlib==0.21.1
cobble==0.1.4
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.2
cookiecutter==2.6.0
crashtest==0.4.1
cryptography==45.0.5
cssselect2==0.8.0
customtkinter==5.2.2
cycler==0.12.1
cymem==2.0.11
darkdetect==0.8.0
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
docstring_parser==0.16
docutils==0.21.2
dulwich==0.22.8
durationpy==0.10
emoji==2.14.1
et_xmlfile==2.0.0
executing==2.2.0
ezdxf==1.4.2
faiss-cpu==1.11.0.post1
fastapi==0.115.14
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
findpython==0.6.3
firebase_admin==7.1.0
flatbuffers==25.2.10
flet==0.28.3
flet-cli==0.28.3
flet-desktop==0.28.3
fonttools==4.58.5
fpdf==1.7.2
fpdf2==2.8.3
frozenlist==1.7.0
fsspec==2025.5.1
ghp-import==2.1.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-secret-manager==2.24.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.3
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.2
humanfriendly==10.0
hyperframe==6.1.0
identify==2.6.12
idna==3.10
imagesize==1.4.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
installer==0.7.0
instructor==1.9.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jaconv==0.4.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keyring==25.6.0
kiwisolver==1.4.8
kubernetes==33.1.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.4.8
language_data==1.3.0
lxml==6.0.0
mammoth==1.9.1
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
mermaid-magic==0.1.4
mermaid-py==0.8.0
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.15
mkdocs-material-extensions==1.3.1
mmh3==5.1.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multimethod==1.12
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==2.0.1
nest-asyncio==1.6.0
networkx==3.5
nlopt==2.9.1
nltk==3.9.1
nodeenv==1.9.1
numpy==2.2.6
oauthlib==3.3.1
olefile==0.47
onnxruntime==1.22.1
openai==1.98.0
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.11.1
oscrypto==1.3.0
overrides==7.7.0
packaging==25.0
paginate==0.5.7
pandas==2.3.1
parso==0.8.4
path==17.1.0
pathspec==0.12.1
pbs-installer==2025.7.23
pdf2image==1.17.0
pefile==2023.2.7
pillow==11.3.0
pip-tools==7.4.1
pkginfo==1.12.1.2
platformdirs==4.3.8
plotly==6.2.0
pluggy==1.6.0
poetry==2.1.3
poetry-core==2.1.3
posthog==5.4.0
pre_commit==4.2.0
preshed==3.0.10
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
PyJWT==2.10.1
pymdown-extensions==10.16
PyMuPDF==1.26.3
pyparsing==3.2.3
pypdf==5.7.0
PyPDF2==3.0.1
PyPika==0.48.9
pypng==0.20220715.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.4.1
pytest-asyncio==1.1.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
python-slugify==8.0.4
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyyaml_env_tag==1.1
qrcode==7.4.2
rank-bm25==0.2.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
repath==0.9.0
reportlab==4.4.2
requests==2.32.4
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
roman-numerals-py==3.1.0
rpds-py==0.26.0
rsa==4.9.1
ruff==0.12.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==5.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
snscrape==0.7.0.20230622
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.41
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
streamlit==1.47.1
streamlit-drawable-canvas==0.9.3
SudachiDict-core==20250515
SudachiDict-full==20250515
SudachiPy==0.6.10
svglib==1.5.1
sympy==1.14.0
tenacity==9.1.2
text-unidecode==1.3
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tldextract==5.3.0
tokenizers==0.21.2
toml==0.10.2
tomlkit==0.13.3
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.1
trimesh==4.6.13
trove-classifiers==2025.5.9.12
typer==0.16.0
types-python-dateutil==2.9.0.20250708
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
typish==1.9.3
tzdata==2025.2
tzlocal==5.3.1
unstructured==0.18.11
unstructured-client==0.41.0
uritemplate==4.2.0
uritools==5.0.0
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.31.2
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wordcloud==1.9.4
wrapt==1.17.2
xhtml2pdf==0.2.17
xlsxwriter==3.2.5
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
